requirements.txt
Django==5.2.4
Pillow==11.3.0
requests==2.32.4
